all lambda functions working

+++ queries.sql
@@ -32,7 +32,7 @@ CREATE TABLE scores
     scoreid     int not null AUTO_INCREMENT,
     playerid    int not null,
     gameid      int not null,
-    score       int not null,
+    score       float not null,
     PRIMARY KEY (scoreid),
     FOREIGN KEY (playerid) REFERENCES players(playerid),
     FOREIGN KEY (gameid) REFERENCES games(gameid)
